feat-504: PR comment: refactor badges into separate files
