+++ web/src/components/Common/Badges/VariantOrLineageText.tsx
@@ -1,0 +1,22 @@
+import { useRecoilValue } from 'recoil'
+import React from 'react'
+import { enablePangolinAtom } from 'src/state/Nomenclature'
+import { clusterPangoLineageMapSelector } from 'src/state/Clusters'
+
+export function VariantOrLineageText({ name, invert }: VariantOrLineageTextProps) {
+  const enablePangolin = useRecoilValue(enablePangolinAtom)
+  const pangoLineageMap = useRecoilValue(clusterPangoLineageMapSelector)
+  const pangoName = pangoLineageMap.get(name) ?? name
+  const showLineageBadge = enablePangolin != invert
+  return <>{showLineageBadge ? pangoName : name}</>
+}
+
+export interface VariantOrLineageTextProps {
+  name: string
+  invert?: boolean
+}
+
+/** Shorter convenience alias for VariantOrLineageText*/
+export function VarOrLinText({ name, invert = false }: VariantOrLineageTextProps) {
+  return <VariantOrLineageText name={name} invert={invert} />
+}
